+++ packages/soda-react/src/utils/isReactNode.ts
@@ -1,0 +1,18 @@
+import { ReactNode, isValidElement } from "react"
+import { isIterable } from "deepsea-tools"
+
+/**
+ * 判断是否为React节点，暂时无法判断 Promise
+ * @param {unknown} value 任意值
+ * @returns {boolean} 是否为React节点
+ */
+export function isReactNode(value: unknown): value is ReactNode {
+    const type = typeof value
+    if (type === "string" || type === "number" || type === "boolean" || type === "undefined" || type === "bigint" || value === null) return true
+    if (type === "function" || type === "symbol") return false
+    if (isIterable(value)) {
+        const newValue = Array.isArray(value) ? value : Array.from(value)
+        return newValue.every(isReactNode)
+    }
+    return isValidElement(value)
+}