+++ packages/deepsea-tools/CHANGELOG.md
@@ -1,5 +1,11 @@
 # deepsea-tools
 
+## 5.31.0
+
+### Minor Changes
+
+- 优化大量内容
+
 ## 5.30.0
 
 ### Minor Changes