+++ packages/soda-heroui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # soda-tanstack-form
 
+## 0.6.5
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.31.0
+    - soda-hooks@6.10.0
+
 ## 0.6.4
 
 ### Patch Changes